+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.1.0-rc.0] - Unreleased
-- Fix csv import when the file contains strings with commas
+### Fixed
+- csv import when the file contains strings with commas.
 
 ## [0.1.0-beta.2] - 2021-01-08
 ### Added