+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.1.0-rc.0] - Unreleased
+## [0.1.0] - 2021-03-08
 ### Added
 - Search filters in the pin weapons section of the optimization page.
 - Weapon levels in YourDeck page.